Added drink view screen- rebuild Home App components for reuse
