+++ src/components/TheNavbar.vue
@@ -1,6 +1,17 @@
 <template>
   <nav
-    class="fixed bottom-0 w-full h-16 left-1/2 -translate-x-1/2 flex items-center justify-center bg-blue-100"
+    class="
+      fixed
+      bottom-0
+      w-full
+      h-16
+      left-1/2
+      -translate-x-1/2
+      flex
+      items-center
+      justify-center
+      bg-blue-100
+    "
   >
     <ul class="flex w-full max-w-sm gap-10 justify-center">
       <navbar-list-item
@@ -17,6 +28,13 @@
       >
         <water-drop-icon class="w-6 h-6 fill-dark"></water-drop-icon>
       </navbar-list-item>
+      <navbar-list-item
+        :class="activeRouterName === 'drinkApp' ? 'shadow-inset-light' : ''"
+        label="Przejdź do strony głównej"
+        path="/drink"
+      >
+        <bottle-icon class="w-6 h-6 fill-dark"></bottle-icon>
+      </navbar-list-item>
 
       <navbar-list-item
         :class="activeRouterName === 'settings' ? 'shadow-inset-light' : ''"
@@ -30,12 +48,15 @@
 </template>
 
 <script setup>
-import { ref } from 'vue';
-import { useRouter } from 'vue-router';
-import ChartIcon from './icons/ChartIcon.vue';
-import SettingsIcon from './icons/SettingsIcon.vue';
-import NavbarListItem from './NavbarListItem.vue';
-import WaterDropIcon from './icons/WaterDropIcon.vue';
+import { ref } from "vue";
+import { useRouter } from "vue-router";
+import NavbarListItem from "./NavbarListItem.vue";
+/* Icons */
+
+import WaterDropIcon from "./icons/WaterDropIcon.vue";
+import ChartIcon from "./icons/ChartIcon.vue";
+import SettingsIcon from "./icons/SettingsIcon.vue";
+import BottleIcon from "./icons/BottleIcon.vue";
 
 const router = useRouter();
 const activeRouterName = ref(router.currentRoute.value.name);
