+++ src/router/index.js
@@ -22,6 +22,21 @@ const router = createRouter({
       },
     },
     {
+      path: '/drink',
+      name: 'drinkApp',
+      component: () => import('../views/DrinkView.vue'),
+      beforeEnter: async (to, from, next) => {
+        if ((await userIsAuth()) && (await checkIfDateExistsInDB())) {
+          next();
+        } else if (
+          (await userIsAuth()) &&
+          (await checkIfDateExistsInDB()) === false
+        ) {
+          next({ name: 'gender' });
+        } else next({ name: 'signin' });
+      },
+    },
+    {
       path: '/settings',
       name: 'settings',
       component: () => import('../views/SettingsView.vue'),
